[WIP] Added x:Name for the user name entry in RegistrationPage
The code works for Handle_Clicked(), but it's WIP becuase need to bind the Text value to the RegistrationPageViewModel in order for the ViewModel to get the Text value.
Also need the view model to get the individual values of the other entries, or just pass the value of the Item object to the view model.
Related to #16.