[WIP] Retrieving a single item from the DynamoDB table It's WIP because there's an exception: ReaditemAsync() Exception = Unable to convert range key value for property SavedTimeStamp
The reason is that Load() loads a single item, so the entire primary key is needed.
See https://stackoverflow.com/a/29664755/4797855
So to get rid of this exception, need to remove/change the range key of the DynamoDB table.

Also remaned the hash parameter in ReadItemAsync() to Id for better code readability.
Also created a button in RegistrationPage.xaml that fires the command RetrievePersonCommand.