Updated Xamarin.Forms to 3.0.0.446417 and other packages except Xamarin.Android.Support Could not update the Xamarin.Android.Support packages due to incompatibility issues:
E.g. Could not install package 'Xamarin.Android.Support.Animated.Vector.Drawable 27.0.2'. You are trying to install this package into a project that targets 'MonoAndroid,Version=v8.0', but the package does not contain any assembly references or content files that are compatible with that framework. For more information, contact the package author.